Use UIAlertController instead of UIAlertView

+++ SeafProvider/DocumentPickerViewController.m
@@ -22,7 +22,7 @@
 
 -(void)prepareForPresentationInMode:(UIDocumentPickerMode)mode
 {
-    Debug("mode=%d", mode);
+    Debug("mode=%lu", (long)mode);
     [SeafGlobal.sharedObject loadAccounts];
     _conns = SeafGlobal.sharedObject.conns;
     self.tableView.separatorStyle = UITableViewCellSeparatorStyleSingleLine;
